requirements.txt
aiohttp>=3.12.14
aiosqlite>=0.21.0
discord.py>=2.5.2
python-dotenv>=1.1.1
pytz>=2024.2
